page_mon/requirements.txt
selenium~=4.27.1
prometheus_client~=0.21.1
python-dotenv~=1.0.1